backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.97.0",
    "fastapi>=0.136.1",
    "greenlet>=3.4.0",
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
